Compression update

+++ header.h
@@ -19,12 +19,12 @@ struct header {
 	int rid;					// Receiver ID
 	int rid2;					// Used for voteOn() error detection and correction
 	int rid3;					// Used for voteOn() error detection and correction
-	BYTE priority;				// Message's pirority
+	int priority;				// Message's pirority
 	int seqNum;					// Message's position in the queue
 	long payloadSize;			// Number of bytes in payload after this header
-	BYTE payloadType;			// mERR, mNONE, mTXT, mAUD, numMsgTypes
-	BYTE encryption;			// ERR, NONE, XOR, VIG, numOfEnc
-	BYTE compression;			// cERR, cNONE, cHUF, cRLE, numCompTypes
+	int payloadType;			// mERR, mNONE, mTXT, mAUD, numMsgTypes
+	int encryption;				// ERR, NONE, XOR, VIG, numOfEnc
+	int compression;			// cERR, cNONE, cHUF, cRLE, numCompTypes
 	long uncompressedLength;	// Uncompressed message length
 	int checksum;				// Checksum for payload error detection
 	bool errorDC;				// Enable or disable error detection and correciton
